src/apitist/generation/templates/requirements.txt
attrs
pytest
python-apitist
tox
